packages/strands-agents/analysis-agent/requirements.txt
strands-agents
bedrock-agentcore
boto3
aws-opentelemetry-distro>=0.10.1
